Rename files for better consistency
