+++ api-rework/invasives/api/serializers/type/subtype/monitoring_treatment_mechanical.py
