+++ api-rework/invasives/api/serializers/type/subtype/treatment_mechanical_terrestrial.py
